+++ src/hooks/useOrderForm.ts
@@ -1,36 +1,87 @@
-import { useState } from 'react'
-import { useSender } from '@/hooks/useSender'
-import { useReceiver } from '@/hooks/useReceiver'
-import { useCardMessage } from '@/hooks/useCardMessage'
+import { useState, useMemo } from 'react'
+import { cardMock } from '@/pages/OrderPage/cardMock'
 
 export function useOrderForm() {
-  const sender = useSender()
-  const { receiver, receiverPhone, quantity } = useReceiver()
-  const card = useCardMessage()
+  const [selectedCard, setSelectedCard] = useState(cardMock[0])
+  const [message, setMessage] = useState(selectedCard.defaultTextMessage)
+  const [sender, setSender] = useState('')
+  const [receiver, setReceiver] = useState('')
+  const [receiverPhone, setReceiverPhone] = useState('')
+  const [quantity, setQuantity] = useState(1)
   const [isSubmitted, setIsSubmitted] = useState(false)
 
+  const senderError = useMemo(() => {
+    if (!sender) return '보내는 사람 이름을 입력해주세요.'
+    return ''
+  }, [sender])
+
+  const receiverError = useMemo(() => {
+    if (!receiver) return '받는 사람 이름을 입력해주세요.'
+    return ''
+  }, [receiver])
+
+  const phoneRegex = /^010\d{8}$/
+  const receiverPhoneError = useMemo(() => {
+    if (!receiverPhone) return '받는 사람 전화번호를 입력해주세요.'
+    if (!phoneRegex.test(receiverPhone))
+      return '올바른 전화번호 형식이 아닙니다.'
+    return ''
+  }, [receiverPhone])
+
+  const messageError = useMemo(() => {
+    if (!message) return '메시지를 입력해주세요.'
+    return ''
+  }, [message])
+
+  const quantityError = useMemo(() => {
+    if (quantity < 1) return '수량은 1개 이상이어야 합니다.'
+    return ''
+  }, [quantity])
+
   const isFormValid =
-    !sender.error &&
-    !receiver.error &&
-    !receiverPhone.error &&
-    !card.error &&
-    !quantity.error
+    !senderError &&
+    !receiverError &&
+    !receiverPhoneError &&
+    !messageError &&
+    !quantityError
 
   const resetForm = () => {
-    sender.set('')
-    receiver.set('')
-    receiverPhone.set('')
-    card.setMessage('')
-    quantity.set(1)
+    setSender('')
+    setReceiver('')
+    setReceiverPhone('')
+    setMessage('')
+    setQuantity(1)
     setIsSubmitted(false)
   }
 
   return {
-    sender,
-    receiver,
-    receiverPhone,
-    quantity,
-    card,
+    sender: {
+      value: sender,
+      set: setSender,
+      error: senderError,
+    },
+    receiver: {
+      value: receiver,
+      set: setReceiver,
+      error: receiverError,
+    },
+    receiverPhone: {
+      value: receiverPhone,
+      set: setReceiverPhone,
+      error: receiverPhoneError,
+    },
+    quantity: {
+      value: quantity,
+      set: setQuantity,
+      error: quantityError,
+    },
+    card: {
+      selectedCard,
+      setSelectedCard,
+      message,
+      setMessage,
+      error: messageError,
+    },
     isSubmitted,
     setIsSubmitted,
     isFormValid,
